# api/openapi/turbovas-v1.yaml
...
openapi: 3.1.0
info:
  title: TurboVAS Native API
  version: 0.1.0-contract
  description: Read-only contract baseline for DB-backed TurboVAS product workflows. Direct listener responses include X-Request-Id for correlation.
servers:
  - url: /api/v1
  - url: http://127.0.0.1:19080/api/v1
    description: Opt-in development direct API listener; requires bearer authentication.
security:
  - operatorSession: []
  - bearerAuth: []
paths:
...
  /cpes:
    get:
      operationId: getCpes
      x-turbovas-direct: true
      x-turbovas-exposure: direct-read
      x-turbovas-maturity: live-read
      x-turbovas-replaces: cpe-catalog-list-read
      x-turbovas-inherited-still-owns: scap-rich-context
      summary: List Security Information CPEs
      description: Searchable CPE intelligence catalog backed by SCAP feed metadata in gvmd/PostgreSQL. This is distinct from observed operating-system and application evidence in raw and scope reports.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: CPE catalog collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CatalogCpeCollection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /cert-bund-advisories:
    get:
      operationId: getCertBundAdvisories
      x-turbovas-direct: true
      x-turbovas-exposure: direct-read
      x-turbovas-maturity: live-read
      x-turbovas-replaces: cert-bund-advisory-list-read
      x-turbovas-inherited-still-owns: cert-advisory-rich-detail-export
      summary: List Security Information CERT-Bund advisories
      description: Searchable CERT-Bund advisory list backed by CERT feed metadata already imported into gvmd/PostgreSQL. Rich advisory detail remains inherited from feed XML/XSLT.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: CERT-Bund advisory collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CertBundAdvisoryCollection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /dfn-cert-advisories:
    get:
      operationId: getDfnCertAdvisories
      x-turbovas-direct: true
      x-turbovas-exposure: direct-read
      x-turbovas-maturity: live-read
      x-turbovas-replaces: dfn-cert-advisory-list-read
      x-turbovas-inherited-still-owns: cert-advisory-rich-detail-export
      summary: List Security Information DFN-CERT advisories
      description: Searchable DFN-CERT advisory list backed by CERT feed metadata already imported into gvmd/PostgreSQL. Rich advisory detail remains inherited from feed XML/XSLT.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: DFN-CERT advisory collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DfnCertAdvisoryCollection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /reports:
    get:
      operationId: getReports
      x-turbovas-direct: true
      x-turbovas-exposure: direct-read
      x-turbovas-maturity: live-read
      x-turbovas-replaces: raw-report-list-read
      x-turbovas-inherited-still-owns: raw-report-generation-xml-export-retention-and-mutations
      summary: List raw scan reports
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Filter'
      responses:
        '200':
          description: Raw report collection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportCollection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '405':
          $ref: '#/components/responses/MethodNotAllowed'
        '413':
          $ref: '#/components/responses/RequestTooLarge'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    operatorSession:
      type: apiKey
      in: cookie
      name: turbovas_session
      description: Authenticated same-origin operator session used by the GSA/gsad browser proxy.
    bearerAuth:
      type: http
      scheme: bearer
      description: Bearer token for the opt-in direct native API development listener.
  responses:
    BadRequest:
      description: Request parameters, path values, or request shape are invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication is missing, invalid, or expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    MethodNotAllowed:
      description: The endpoint supports only the documented HTTP method.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    RequestTooLarge:
      description: The direct native API request exceeded the accepted size limits.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    TooManyRequests:
      description: The direct native API listener is handling too many concurrent requests.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: The native API could not complete the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    Sort:
      name: sort
      in: query
      schema:
        type: string
    Filter:
      name: filter
      in: query
      schema:
        type: string
        maxLength: 4096
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
    PageInfo:
      type: object
      required: [page, page_size, total]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        sort:
          type: string
        filter:
          type: string
    ReportCollection:
      type: object
      required: [page, items]
      properties:
        page:
          $ref: '#/components/schemas/PageInfo'
        items:
          type: array
          items:
            $ref: '#/components/schemas/Report'
...
    CatalogCpeCollection:
      type: object
      required: [page, items]
      properties:
        page:
          $ref: '#/components/schemas/PageInfo'
        items:
          type: array
          items:
            $ref: '#/components/schemas/CatalogCpe'
    DfnCertAdvisoryCollection:
      type: object
      required: [page, items]
      properties:
        page:
          $ref: '#/components/schemas/PageInfo'
        items:
          type: array
          items:
            $ref: '#/components/schemas/DfnCertAdvisory'
    CertBundAdvisoryCollection:
      type: object
      required: [page, items]
      properties:
        page:
          $ref: '#/components/schemas/PageInfo'
        items:
          type: array
          items:
            $ref: '#/components/schemas/CertBundAdvisory'
...
    Report:
      type: object
      required: [id, name, owner, status, result_count, vulnerability_count, host_count, cve_count, severity, max_severity]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        owner:
          $ref: '#/components/schemas/ReportOwner'
        status:
          type: string
        task:
          $ref: '#/components/schemas/ReportReference'
        target:
          $ref: '#/components/schemas/ReportReference'
        creation_time:
          type: string
          format: date-time
        modification_time:
          type: string
          format: date-time
        scan_start:
          type: string
          format: date-time
        scan_end:
          type: string
          format: date-time
        severity:
          $ref: '#/components/schemas/ReportSeverityCounts'
        max_severity:
          type: number
        result_count:
          type: integer
        vulnerability_count:
          type: integer
        host_count:
          type: integer
        cve_count:
          type: integer
        user_tags:
          type: array
          description: Active user tags attached directly to this report. Generic report/result tag-resource enumeration remains outside the native tag-resource API allowlist.
          items:
            $ref: '#/components/schemas/ReportUserTag'
    ReportOwner:
      type: object
      required: [name]
      properties:
        name:
          type: string
    ReportReference:
      type: object
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    ReportSeverityCounts:
      type: object
      required: [critical, high, medium, low, log, false_positive]
      properties:
        critical:
          type: integer
        high:
          type: integer
        medium:
          type: integer
        low:
          type: integer
        log:
          type: integer
        false_positive:
          type: integer
    ReportUserTag:
      type: object
      required: [id, name, value, comment]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        value:
          type: string
        comment:
          type: string
...
    CatalogCpeCve:
      type: object
      required: [id, severity]
      properties:
        id:
          type: string
          pattern: '^CVE-[0-9]{4}-[0-9]{4,}$'
        severity:
          type: number
    CatalogCpe:
      type: object
      required: [id, name, deprecated, cve_refs]
      properties:
        id:
          type: string
          description: CPE URI, for example cpe:/a:vendor:product:version.
        name:
          type: string
        comment:
          type: string
        title:
          type: string
        cpe_name_id:
          type: string
        deprecated:
          type: boolean
        deprecated_by:
          type: string
        severity:
          type: number
          nullable: true
        cve_refs:
          type: integer
        cves:
          type: array
          items:
            $ref: '#/components/schemas/CatalogCpeCve'
        created_at:
          type: string
          format: date-time
        modified_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DfnCertAdvisory:
      type: object
      required: [id, name, cve_refs]
      properties:
        id:
          type: string
        name:
          type: string
        comment:
          type: string
        title:
          type: string
        summary:
          type: string
        severity:
          type: number
        cve_refs:
          type: integer
        cves:
          type: array
          items:
            type: string
            pattern: '^CVE-[0-9]{4}-[0-9]{4,}$'
        created_at:
          type: string
          format: date-time
        modified_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CertBundAdvisory:
      type: object
      required: [id, name, cve_refs]
      properties:
        id:
          type: string
        name:
          type: string
        comment:
          type: string
        title:
          type: string
        summary:
          type: string
        severity:
          type: number
        cve_refs:
          type: integer
        cves:
          type: array
          items:
            type: string
            pattern: '^CVE-[0-9]{4}-[0-9]{4,}$'
        created_at:
          type: string
          format: date-time
        modified_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...